Blog level 6: 1) overriding methods equals and hashcode for Comment, changed them for Post and Tag (now all fields are taking part in both methods) 2) Post: adding cascade type all for tags set. Also initialise list of comments to fix org.hibernate.HibernateException: A collection with cascade=”all-delete-orphan” was no longer referenced by the owning entity instance: com.children 3) PostController: get all methods noe returns dto with comments and tags 4)PostControllerTest: was updated by adding mock TagService 5) PostServiceTest: delete method covered by test 6) Tag: name field now is unique to avoid adding the same tags to db 7) TagRepository, TagService, TagServiceTest: the logic for save and delete was changed mainly to dealt with tag's name instead of id, tests were adding too

+++ src/main/java/com/luxcampus/Blog/service/TagService.java
@@ -8,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 
@@ -20,43 +20,43 @@ public class TagService implements TagServiceInterface {
     private final PostRepository postRepository;
     private final TagRepository tagRepository;
 
-
     @Transactional
     @Override
     public void addTagToPost(Tag tag, Long postId) {
+        //check if tag from request is already in db
+        Tag tagFromDB = tagRepository.findByName(tag.getName());
         Post post = postRepository.getById(postId);
-        //check if this tag is already exist in post
-        Set<Tag> tagsInPost = post.getTags();
-        for (Tag tagInPost : tagsInPost) {
-            if(tagInPost.getName().equals(tag.getName())){
-                return;
+        if (tagFromDB==null){
+            //it is new tag, it should be saved in db
+            post.getTags().add(tag);
+            tag.getPosts().add(post);
+            tagRepository.save(tag);
+        }else {
+            //check if this tag is already exist in post
+            Set<Tag> tagsInPost = post.getTags();
+            for (Tag tagInPost : tagsInPost) {
+                if(tagInPost.getName().equals(tagFromDB.getName())){
+                    return;
+                }
             }
+            //this post has not such tagFromDB, it should be added
+            post.getTags().add(tagFromDB);
+            tagFromDB.getPosts().add(post);
         }
-        //check if this a new tag
-        List<Tag> tags = tagRepository.findAll();
-        for (Tag existTag : tags) {
-            if (tag.getName().equals(existTag.getName())) {
-                post.getTags().add(existTag);
-                existTag.getPosts().add(post);
-                return;
-            }
-        }
-        post.getTags().add(tag);
-        tag.getPosts().add(post);
-        tagRepository.save(tag);
     }
 
     @Override
     @Transactional
     public void deleteTagFromPost(Long id) {
-        Tag tag = tagRepository.getById(id);
-
-        List<Post> postList = postRepository.findAll();
-        for (Post post : postList) {
-            Set<Tag> tagSet = post.getTags();
-            HashSet set = new HashSet(tagSet);
-            set.remove(tag);
+        Optional<Tag> optionalTag = tagRepository.findById(id);
+        if (optionalTag.isPresent()) {
+            List<Post> postList = postRepository.findAll();
+            Tag tag = optionalTag.get();
+            for (Post post : postList) {
+                Set<Tag> tagSet = post.getTags();
+                tagSet.remove(tag);
+            }
+            tagRepository.delete(tag);
         }
-        tagRepository.delete(tag);
     }
 }
